Clang-format reodered shader includes that messed up #defines.

+++ backends/graphics/sopengl/src/sigma/opengl/texture.cpp
@@ -23,7 +23,7 @@ namespace opengl {
     }
 
     texture::texture(graphics::texture_data data)
-        : texture(internal_format::RGBA8, data.size, data.pixels)
+        : texture(internal_format::RGBA32F, data.size, data.pixels)
     {
     }
